requirements.txt
Django>=2.2.1
django-bootstrap4>=0.0.8
django-bootstrap-datepicker-plus>=3.0.5
psycopg2-binary>=2.8.2
